backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic
python-docx
python-pptx
openai
python-multipart